Add isValidDate prop to DateTimePicker
isValidDate prop gives the possibility to disable dates in the calendar if the users are not allowed to select them, e.g. dates in the future. The function shall return true for selectable dates, and false for disabled ones.

Default value for isValidDate prop is a function that always returns true.